+++ src/components/Admin/Brand/BrandTable.tsx
@@ -18,6 +18,7 @@ import { usePagination } from "../../../hooks/usePagination";
 import { Box, TextField } from "@mui/material";
 import SearchIcon from "@mui/icons-material/Search";
 import { useSearch } from "../../../hooks/useSearch";
+import useHasAccess from "../../../hooks/useHasAccess";
 
 interface Column {
   id: "index" | "code" | "name" | "createAt" | "actions" | "image";
@@ -47,18 +48,41 @@ function BrandTable() {
   } = useContext(BrandContext) as brandContextType;
 
   const [currentPage, setCurrentPage] = useState(1);
+
+  const { userHasAccess: accessList } = useHasAccess("action_brand_admin_list");
+  const { userHasAccess: accessDelete } = useHasAccess(
+    "action_brand_admin_delete"
+  );
+  const { userHasAccess: accessEdit } = useHasAccess(
+    "action_brand_admin_update"
+  );
+
   const [deleteBrand] = useDeleteBrandMutation();
 
   const deleteBrandHandler = async (id: string) => {
-    deleteBrand(id);
-    await deleteBrand(id).then(() => {
-      toast.success("Delete Brand Is Success");
-      refetchBrands();
+    if (accessDelete) {
+      deleteBrand(id);
+      await deleteBrand(id).then(() => {
+        toast.success("Delete Brand Is Success");
+        refetchBrands();
+        setShowDeleteModal(false);
+      });
+    } else {
+      toast.error("You Havent Access Delete Brand");
       setShowDeleteModal(false);
-    });
+    }
   };
 
-  const pageSize = 8;
+  const editBrandHandler = (id: string) => {
+    if (accessEdit) {
+      setOpenEditModal(true);
+      setEditBrandId(id);
+    } else {
+      toast.error("You Havent Access Edit Brand");
+    }
+  };
+
+  const pageSize = 9;
   const {} = usePagination(currentPage, pageSize);
   const totalPages = Math.ceil(total / pageSize);
   const changePageHandler = (id: number) => {
@@ -78,7 +102,7 @@ function BrandTable() {
   };
 
   return (
-    <div className="lg:col-span-8 col-span-12 lg:order-1 order-2 rounded-xl bg-white mt-3">
+    <div className="lg:col-span-8 col-span-12 lg:order-1 order-2 rounded-xl bg-white m-3">
       <Paper
         sx={{
           width: "100%",
@@ -88,7 +112,7 @@ function BrandTable() {
           borderRadius: "12px",
         }}
       >
-        <div className="h-8  md:mx-8 mb-4">
+        <div className="h-8 md:mx-8 mb-4">
           <Box sx={{ display: "flex", alignItems: "flex-end" }}>
             <SearchIcon
               sx={{ color: "action.active", mr: 1, my: 0.5 }}
@@ -104,81 +128,89 @@ function BrandTable() {
             />
           </Box>
         </div>
-        <TableContainer sx={{ minHeight: 500 }}>
-          <Table stickyHeader aria-label="sticky table">
-            {brandsLoading ? (
-              <Spinner />
-            ) : (
-              <>
-                <TableHead>
-                  <TableRow>
-                    {columns.map((column) => (
-                      <TableCell key={column.id} align="center">
-                        {column.label}
-                      </TableCell>
-                    ))}
-                  </TableRow>
-                </TableHead>
-                <TableBody>
-                  {brands?.length ? (
-                    brands?.map((row: any, index: any) => (
-                      <TableRow key={row.id}>
-                        <TableCell align="center">{index + 1}</TableCell>
-                        <TableCell align="center">{row.name}</TableCell>
-                        <TableCell align="center">{row.code}</TableCell>
-                        <TableCell
-                          align="center"
-                          style={{ display: "flex", justifyContent: "center" }}
-                        >
-                          <img
-                            src={`http://127.0.0.1:6060/${row.fileUrl}`}
-                            className="object-contain max-h-6 max-w-6"
-                          />
+        <TableContainer sx={{ minHeight: 615 }}>
+          {accessList ? (
+            <Table stickyHeader aria-label="sticky table">
+              {brandsLoading ? (
+                <Spinner />
+              ) : (
+                <>
+                  <TableHead>
+                    <TableRow>
+                      {columns.map((column) => (
+                        <TableCell key={column.id} align="center">
+                          {column.label}
                         </TableCell>
-                        <TableCell
-                          style={{ whiteSpace: "nowrap" }}
-                          align="center"
-                        >
-                          {row.createdAt.slice(0, 10)}
-                        </TableCell>
-                        <TableCell
-                          align="center"
-                          sx={{
-                            display: "flex",
-                            justifyContent: "center",
-                            mt: "6px",
-                          }}
-                        >
-                          <FaTrash
-                            className="text-red mr-3"
-                            onClick={() => {
-                              setDeleteBrandId(row.id);
-                              setShowDeleteModal(true);
+                      ))}
+                    </TableRow>
+                  </TableHead>
+                  <TableBody>
+                    {brands?.length ? (
+                      brands?.map((row: any, index: any) => (
+                        <TableRow key={row.id}>
+                          <TableCell align="center">{index + 1}</TableCell>
+                          <TableCell align="center">{row.name}</TableCell>
+                          <TableCell align="center">{row.code}</TableCell>
+                          <TableCell
+                            align="center"
+                            style={{
+                              display: "flex",
+                              justifyContent: "center",
                             }}
-                          />
-                          <FaPen
-                            className="text-orange-500"
-                            onClick={() => {
-                              setOpenEditModal(true);
-                              setEditBrandId(row.id);
+                          >
+                            <img
+                              src={row.fileUrl}
+                              className="object-contain max-h-6 max-w-6"
+                            />
+                          </TableCell>
+                          <TableCell
+                            style={{ whiteSpace: "nowrap" }}
+                            align="center"
+                          >
+                            {row.createdAt.slice(0, 10)}
+                          </TableCell>
+                          <TableCell
+                            align="center"
+                            sx={{
+                              display: "flex",
+                              justifyContent: "center",
+                              mt: "6px",
                             }}
-                          />
+                          >
+                            <FaTrash
+                              className="text-red mr-3"
+                              onClick={() => {
+                                setDeleteBrandId(row.id);
+                                setShowDeleteModal(true);
+                              }}
+                            />
+                            <FaPen
+                              className="text-orange-500"
+                              onClick={() => editBrandHandler(row.id)}
+                            />
+                          </TableCell>
+                        </TableRow>
+                      ))
+                    ) : (
+                      <TableRow>
+                        <TableCell colSpan={7}>
+                          <p className="text-xl text-center font-semibold">
+                            We couldn't find a brand with these specifications.
+                          </p>
                         </TableCell>
                       </TableRow>
-                    ))
-                  ) : (
-                    <TableRow>
-                      <TableCell colSpan={7}>
-                        <p className="text-xl text-center font-semibold">
-                          We couldn't find a brand with these specifications.
-                        </p>
-                      </TableCell>
-                    </TableRow>
-                  )}
-                </TableBody>
-              </>
-            )}
-          </Table>
+                    )}
+                  </TableBody>
+                </>
+              )}
+            </Table>
+          ) : (
+            <Box sx={{ marginTop: "100px" }}>
+              <h1 className="text-3xl font-bold  flex justify-center items-center ">
+                You Havent Access Brand List
+              </h1>
+            </Box>
+          )}
         </TableContainer>
 
         {totalPages > 1 && (
